fix warnings and add -Werror

+++ waze/address_search/address_search_dlg.c
@@ -157,7 +157,7 @@ static void on_address_resolved( void*                context,
 
    SsdWidget list_cont = (SsdWidget)context;
    SsdWidget list;
-   int       i;
+   intptr_t       i;
 
    s_searching = FALSE;
 
@@ -386,7 +386,7 @@ static int get_selected_list_item()
    list = ssd_widget_get( dlg, ASD_RC_LIST_NAME);
   waze_assert(list);
 
-   return (int)ssd_list_selected_value( list);
+   return (intptr_t)ssd_list_selected_value( list);
 }
 
 
@@ -416,7 +416,7 @@ static void on_option_selected(  BOOL              made_selection,
    {
       case cm_navigate:
          do_nav = TRUE;
-         // Roll down...
+         /* fall through */
       case cm_show:
          close = navigate( do_nav);
          break;
